2023/2/18

+++ User/FTU.h
@@ -36,7 +36,8 @@ typedef struct
   
   uint16_t RxSize; 
   uint16_t TxSize; 
-  uint16_t RxCount;
+  
+  uint32_t RxCount;
   uint32_t RxTime;
   
   uint8_t RxEndFlag;
